Initial code for storing histogram data
This is for storing statistics for wellknown domains which will be
written out as another parquet file. Right now we can create a parquet
schema based on the struct, but writing out the actual file is still to
be completed.

While here create common function for turning a list of DNS labels to
the reversed layout used in both dnstap arrow data as well as the new
histogram struct.