Use DAWG data for filtering well-known domains
Idea and concept testing by @morkrost. It saves us RAM compared to
storing the equivalent data in a map. Lookups are slower compared to a
map but we will have to see if that becomes a problem or not.

While here add a basic benchmark test for the isKnown() method so we can
see impact of changes to it. It required the existance of a
"well-known-domains.dawg" in the directory, it can created with
https://github.com/dnstapir/dtm-dawg-maker.