Start writing timestamped parquet files
Up until this point it has been OK overwriting the same files over and
over for testing purposes, but now it is time to actually start creating
unique files.

We create uniqueness based on a timestamp and will now store the files
in a persistent location by default (/var/lib/dtm).