Remove dnstap output writing
This code was relevant when this tool was still called "dta" and was
supposed to act as a filter between a dnstap writer and a dnstap reader.
Now that it has taken on the "minimiser" role there is no need to pass
on modified dnstap data to some other party.