Add LRU cache for "seen" domain names
This is to keep track of domains not on the well-known domain list but
we have seen since we started, so we try to only send new_qname events
once for each name.

We use a cache of bounded size instead of something like a map to not
grow indefinitely.