Back out dns related data from arrow structure.
Trying to print out the histogram data resulted in:
```
msg="writeSession: unable to create parquet writer: not implemented: support for DENSE_UNION""
```

So trying to use a union for storing the rdata as the spec requested
meant we are not able to write out parquet data about it. The plan now
is to not extract a bunch of fields from the dns data, instead we will
include the wire-bytes as is.