Stop using arrow for dnstap session data
It is easier to make sure we get the same type of parquet output if we
use the same struct based library as we already do for histogram data.
Since we have currently landed on parquet as the output format from dtm
we are not really using any arrow-specific features other than parquet
writing (e.g. https://arrow.apache.org/use_cases/) so skip the overhead
of using arrow for now.