Support suffix matches of well-known domains
A suffix is entered into the DAWG file by using a leading dot, so
"example.com." will only match exactly that name, while ".example.com."
will match anything below those labels.

Since the labels added to well-known histogram data will be the same for
"example.com." and ".example.com." this commit also adds a
"suffix_match" bool to the histogram struct. This way you can know if the
counters you are looking at is a result of exactly that name or the
combined lookups of stuff below that name.