Migrate to cobra and viper for flags and config
This way we use the same tools as other go code that is appearing in the
project.

Stuff was initialized with:
```
cobra-cli init --viper
cobra-cli add run
```

The main code has been moved from "main" to a "runner" package so we
can import and call it from the cobra-managed "cmd" package.