Rework shutdown of MQTT publishing
Before this the process could hang on shutdown if the MQTT bus was
unreachable for some reason (like expired client certificates).

In this case the mqttPubCh channel would fill up, blocking the channel
write in newQnamePublisher, which in turn made the go routine never exit
even if the input channel was closed.

Now we either write to mqttPubCh or notice that the MQTT-related context
has been cancelled.