Make sure we always close mqttPubCh
The process was not shutting down properly and this was because we would
fail to close mqttPubCh if we didnt happen to be waiting on a write to
mqttPubCh at the moment the paho context was cancelled.

This would in turn make runAutoPaho() block forever trying to read from
mqttPubCh.

Now we always close mqttPubCh.