Pseudonymise IPv4 addresses to native IPv4
When working on the previous commit I noticed that even if the input
IPv4 addresses printed as a normal address, e.g. "127.0.0.1", the
resulting pseudonymised addresses printed as "::ffff:1.2.3.4". This is
because the cryptopan lib returns data via net.IPv4() which always wraps
the address in the 16-byte form.

Now we Unmap() the anonymised address prior to returning it so we get
"native" IPv4 addresses instead.