Make setSessionLabels() match setHistogramLabels()
This code duplication is annoying but it comes down to having different
struct types for histogram and session data which both have the label
fields in them separately. This is because writing out parquet files do
not work with the labels in a separate embedded struct.

While here also duplicate the test so we dont accidentally break
setSessionLabels().