Introduce memoization for cryptopan addresses
Looking at CPU profiling of dtm in pyroscope shows that one of the
functions we spend the most time on is crypto/aes.encryptBlockGo. This
is called as a result of the cryptopan.Anonymize() function.

In order to spend less time on recurring aes encryption try storing the
result in a LRU cache. If the cryptopan instance is updated at runtime
we also replace the cache.

Introduce some prometheus counters around the cache so we can track hits
as well as evictions.