Make cryptopan cache store unmapped IPv4
While adding some logging to the tests showing keys/values in the
cache I noticed that the pseudonymised IPv4 addresses would be stored as
IPv4-mapped IPv6 address, e.g.:
```
cache key: 198.51.100.20, value: ::ffff:185.204.164.235
```

While the address returned from the function was being fixed via the
existing call to .Unmap() it seems better to be consistent, so move the
.Unmap() to right after the address has been calculated. Now addresses
added to the cache will look the same as the one being returned.

While here add tests veryfing the entry in the cache is the same as the
one being returned.