Break out data collection from runMinimiser
Up until this point we have interrupted the minimiser dnstap parsing
loop at regular intervals passing along collected data to the parquet
writers.

Now we no longer do the ticker based interruption in the minimiser loop
but instead do this in a new goroutine. This is another step in making
it possible to run multiple instances of the minimiser loop.