Modify --disable-histogram-sender
Make it behave more like --disable-mqtt. This way http specific flags
are only needed if the histogram sender is not disabled.

While here also add some more logging for worker goroutines starting.
Now that some of them can be disabled it makes sense to be able to
tell directly at startup what is running and what is not.

While here also remove the confusing "waiting for minimiser workers to
exit" log message. While the message is strictly true, looking at the
logs makes it appear as if the program is shutting down which it is not.