Make minimiser workers fight less over lock
Instead of having all workers compete for the same central lock for each
dnstap packet in order to fill in the well-known-domain histogram data,
have them take a RO-lock instead and send the resulting update to a
central goroutine (dataCollector()) which will update the wkd data and
only needs to take a write lock when the rotation timer kicks in.

Because there is no longer a central write lock around modificatins to
the wkd data this however means that an update generated by a worker
might be out of date at the time it is picked up by the dataCollector.
Because of this we also need to check if the dawg file we are currently
using is the same as it was when the update was created. For now we only
use the modification time of the dawg file to keep track of this.

In cases where the update is outdated we implement retries by doing an
additional lookup to verify the name is still considered well known.
This is combined with a retry counter in the update to eventually
discard it if there is somehow an endless loop going on.

Also close the dawg finder instance in histogram writer after it has
been used for lookups if it has been rotated out.

Update tests to use the new functions.