Always call wg.Add(1) before starting goroutine
While it felt prettier to do both wg.Add(1) and wg.Done() inside the
goroutine functions this is inherently causing a race between starting
the goroutine and the later call to wg.Wait().

While things seems to have worked fine this way lets do it correctly
everywhere to not lead by poor example for future work.