Fix revive nit
```
pkg/runner/runner.go:1872:9: indent-error-flow: if block ends with a return statement, so drop this else and outdent its block (revive)
```

The removal of the else-block almost caused accidental reuse of an "ok"
variable. This could cause a bug when caching is not enabled. Instead
replace the previously else-scoped "ok" with a "cacheHit" variable
to more easily keep them separated. Seems more robust than resetting
"ok" to false efter the if-statement.

While here add tests that would catch similar reuse in the future.