Make client cert reloading dynamic
Instead of requiring a restart of edm to pick up new client certificates
for MQTT and HTTP aggregate sending we now use fsnotify monitoring to
dynamically reload a certificate if it is updated on disk.

This also required modification of registerFSWatcher() as it would
previously only support a single callback function for a given filename.
Because MQTT and HTTP aggregate sending may use the same certificate
file we now support the assignment of multiple callbacks for a given
file.