Add persistence to MQTT messages
Now we can survive a process restart without losing messages sitting in
the queue.

This also needed some setup for the autopaho logging since we no longer
get feedback directly at the Publish() call if things went well or not
(we only know if we managed to add to to the local queue).