Improve paho logging and debug logging in general
Make it so we can tag messages from paho internal logging in the slog
messages, so we can now what part of paho created the message.

This also resulted in using the slog Debug level properly. edm will now
set the slog loglevel to debug if the --debug flag is passed. This also
resulted in some cleanup of the more verbose debug messages left in the code
so it is actuallu possible to run edm on a real system with --debug
enabled without generating massive amount of log messages.