Manage configuration in a struct (#46)
* Manage configuration in a struct

Reading config via viper.GetString("foo"), viper.GetInt("bar") etc is
error prone as the string being looked for can easily be misspelled
which will not be noticed at compile time. Now we unmarshal the viper
config to a config struct instead and this way a program that tries to
use an unknown struct field will not even compile which is nice.

This way we can also more easily do validation of the contents of
the config which has been added with the validator package.

* Fix typo in struct field

* Add some validation to conf fields