Stop requiring an input flag (#58)
Since we are now moving in the direction of using a config file we want
to be able to control inputs from the config instead.

There is already validation for the resulting conf struct to catch duplicate
inputs so just remove the requirement from the flags. While here replace
"exluded_with_all" with "exluded_with", a field must no be set if _any_ of
the other fields are set.